Internal: split sycalls tests between Cairo 0 & 1
