+++ tests/integration/main.rs
@@ -1,5 +1,6 @@
 mod common;
 mod declare_txn_tests;
+mod deprecated_syscalls_tests;
 mod os;
 mod pie;
 mod sharp;
